backend/requirements.txt
fastapi
uvicorn
litellm
python-dotenv
asyncio
httpx
